[refactor]: 	- move down the navigation from App.js to Root.js view; 	- extract auxiliary function from useEffect in Chat.js and move to helpers/unshiftMsgs.js 	- add comments
